fix: :bug: findAnnoucement

+++ src/services/annoucement/deleteAnnoucement.services.ts
@@ -2,14 +2,21 @@ import { AppDataSource } from "../../data-source";
 import { Annoucements } from "../../entities/annoucements.entity";
 import { AppError } from "../../errors";
 
-const deleteAnnouncementsService = async(id: string, userId: string) => {
+const deleteAnnouncementsService = async (id: string, userId: string) => {
   const annoucementsRepository = AppDataSource.getRepository(Annoucements);
 
-  const findAnnoucement = await annoucementsRepository.findOneOrFail({where: {annoucementId:id}});
+  const findAnnoucement = await annoucementsRepository.findOne({
+    where: { annoucementId: id },
+    relations: { user: true },
+  });
 
-  if(findAnnoucement.user.id !== userId){
+  if (!findAnnoucement) {
+    throw new AppError("Annoucement not found", 404);
+  }
+
+  if (findAnnoucement.user.id !== userId) {
     throw new AppError("Unauthorized", 401);
-  } 
+  }
 
   await annoucementsRepository.delete(id);
